Removing all old optimizer logic that relies on V2 protocol-owned liquidity pool
Adding instead a more segregated logic that has simpler NFT proposal structure for auctions, supports concentrated liquidity.

The new FutarchyOracle supports a random failing chance for proposals. Easy upgradeability.

Offloads the logic of creating the conditional pools to the proposer -- the contracts only validate that the pools are set correctly.